requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
executing==2.2.0
frozenlist==1.7.0
gevent==25.5.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
gspread==6.2.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
magic-filter==1.0.12
matplotlib-inline==0.1.7
multidict==6.6.4
oauthlib==3.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
stack-data==0.6.3
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
websocket==0.2.1
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2